+++ NEWS.md
@@ -1,4 +1,4 @@
-## rappdirs 0.3.1.9000
+# rappdirs 0.3.2
 
 * `user_data_dir()`, `use_cache_dir()` and `use_config_dir()` now respect
   `R_USER_DATA_DIR`, `R_USER_CACHE_DIR` and `R_USER_CONFIG_DIR` env vars
@@ -6,11 +6,11 @@
 
 * No longer uses methods package.
 
-## rappdirs 0.3.1
+# rappdirs 0.3.1
 
 Minor R CMD check and test fixes.
 
-## rappdirs 0.3.0
+# rappdirs 0.3.0
 
 * first CRAN release
 * `xxx_dir()` functions only use version when appname is not null